Portability fix.
	Partially fix portability problem due to wrong cast
	of size_t to unsigned:
	* mln/histo/data.hh,
	* mln/debug/println_with_border.spe.hh,
	* mln/core/image/translate_image.hh,
	* mln/draw/graph.hh,
	* mln/level/all.hh,
	* mln/accu/count_adjacent_vertices.hh,
	* mln/accu/count.hh,
	* mln/accu/min_h.hh,
	* mln/accu/rank.hh,
	* mln/accu/max_h.hh,
	* mln/accu/median_alt.hh,
	* mln/accu/median_h.hh,
	* mln/accu/volume.hh,
	* mln/accu/height.hh,
	* mln/border/fill.hh,
	* mln/border/duplicate.hh,
	* mln/border/mirror.hh,
	* mln/geom/resize.hh,
	* mln/morpho/closing_volume.hh,
	* mln/morpho/closing_area.hh,
	* mln/morpho/closing_height.hh,
	* mln/morpho/opening_area_on_vertices.hh,
	* mln/morpho/opening_volume.hh,
	* mln/morpho/closing_area_on_vertices.hh,
	* mln/morpho/opening_height.hh,
	* mln/morpho/opening_area.hh: Remove wrong use of std::size_t.


git-svn-id: https://svn.lrde.epita.fr/svn/oln/branches/cleanup-2008@2742 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/draw/graph.hh
@@ -95,7 +95,7 @@ namespace mln
       // Draw the background.
       level::fill(ima, 0);
       // Draw the lines (edges).
-      for (size_t l = 0; l < pg.nedges(); ++l)
+      for (unsigned l = 0; l < pg.nedges(); ++l)
 	line (exact(ima),
 	      // FIXME: Too low-level.  See similar remarks
 	      // in mln/core/image/graph_image.hh
@@ -103,7 +103,7 @@ namespace mln
 	      pg.gr_->vertex_data(pg.gr_->edge(l).v2()),
 	      edge_v);
       // Draw the points (vertices).
-      for (size_t p = 0; p < pg.nsites(); ++p)
+      for (unsigned p = 0; p < pg.nsites(); ++p)
  	exact(ima)(pg.gr_->vertex_data(p)) = vertex_v;
     }
 
@@ -116,10 +116,10 @@ namespace mln
       // Draw the background.
       level::fill(ima, 0);
       // Draw the lines (edges).
-      for (size_t l = 0; l < gi.domain().nedges(); ++l)
+      for (unsigned l = 0; l < gi.domain().nedges(); ++l)
 	line (exact(ima), gi.vertex1(l), gi.vertex2(l), edge_v);
       // Draw the points (vertices).
-      for (size_t p = 0; p < gi.domain().nvertices(); ++p)
+      for (unsigned p = 0; p < gi.domain().nvertices(); ++p)
  	exact(ima)(gi.domain().point_from_id(p)) = gi.vertex_values()[p];
     }
